feat: split post container into pieces
